+++ umongo/document.py
@@ -73,7 +73,7 @@ class DocumentOpts:
     is_child             no                     Document inherit of a non-abstract document
     base_schema_cls      yes                    Base Schema class to use
     indexes              yes                    List of custom indexes
-    children             no                     List of Document inheriting this one
+    offspring            no                     List of Documents inheriting this one
     ==================== ====================== ===========
 
     """
@@ -88,12 +88,12 @@ class DocumentOpts:
                 'is_child={self.is_child}, '
                 'base_schema_cls={self.base_schema_cls}, '
                 'indexes={self.indexes}, '
-                'children={self.children})>'
+                'offspring={self.offspring})>'
                 .format(ClassName=self.__class__.__name__, self=self))
 
     def __init__(self, instance, template, collection_name=None, abstract=False,
                  allow_inheritance=None, base_schema_cls=Schema, indexes=None,
-                 is_child=False, children=None):
+                 is_child=False, offspring=None):
         self.instance = instance
         self.template = template
         self.collection_name = collection_name if not abstract else None
@@ -102,7 +102,7 @@ class DocumentOpts:
         self.base_schema_cls = base_schema_cls
         self.indexes = indexes or []
         self.is_child = is_child
-        self.children = set(children) if children else set()
+        self.offspring = set(offspring) if offspring else set()
         if self.abstract and not self.allow_inheritance:
             raise DocumentDefinitionError("Abstract document cannot disable inheritance")